## Description
Fix calculation of checking for current page size so that
parquet page sizes don't keep reducing with increasing number 
of compressed pages buffered in memory. Small parquet pages
can reduce parquet reader performance by forcing decoding
and decompression on smaller chunks of data.

```
Benchmark                        (benchmarkFileFormat)  (compression)  (dataSet)   Mode  Cnt  Before score    After score
BenchmarkHiveFileFormat.read   TRINO_OPTIMIZED_PARQUET         SNAPPY   LINEITEM  thrpt   20  15.767 ± 0.155  16.652 ± 0.113  ops/s

Before
read  LINEITEM  SNAPPY  TRINO_OPTIMIZED_PARQUET   788.4MB/s ±  7931.1kB/s ( 0.98%) (N = 20, α = 99.9%)

After
read  LINEITEM  SNAPPY  TRINO_OPTIMIZED_PARQUET   832.7MB/s ±  5808.6kB/s ( 0.68%) (N = 20, α = 99.9%)
```
E.g. page sizes before
```
Column: shipmode
--------------------------------------------------------------------------------
  page   type  enc  count   avg size   size       rows     nulls   min / max
  0-D    dict  Z _  7       8.29 B     58 B
  0-1    data  Z R  126991  0.38 B     46.761 kB
  0-2    data  Z R  123519  0.38 B     45.483 kB
  0-3    data  Z R  116706  0.38 B     42.977 kB
  0-4    data  Z R  110477  0.38 B     40.683 kB
  0-5    data  Z R  105683  0.38 B     38.918 kB
  0-6    data  Z R  102963  0.38 B     37.917 kB
  0-7    data  Z R  96078   0.38 B     35.381 kB
  0-8    data  Z R  92639   0.38 B     34.114 kB
  0-9    data  Z R  89242   0.38 B     32.866 kB
  0-10   data  Z R  85820   0.38 B     31.605 kB
  0-11   data  Z R  78934   0.38 B     29.069 kB
  0-12   data  Z R  75509   0.38 B     27.809 kB
  0-13   data  Z R  72057   0.38 B     26.539 kB
  0-14   data  Z R  68657   0.38 B     25.288 kB
  0-15   data  Z R  68668   0.38 B     25.291 kB
  0-16   data  Z R  65214   0.38 B     24.019 kB
  0-17   data  Z R  61749   0.38 B     22.743 kB
  0-18   data  Z R  58357   0.38 B     21.494 kB
  0-19   data  Z R  54919   0.38 B     20.228 kB
  0-20   data  Z R  51487   0.38 B     18.965 kB
  0-21   data  Z R  51485   0.38 B     18.965 kB
  0-22   data  Z R  48028   0.38 B     17.692 kB
  0-23   data  Z R  44600   0.38 B     16.429 kB
  0-24   data  Z R  44605   0.38 B     16.432 kB
  0-25   data  Z R  41188   0.38 B     15.174 kB
  0-26   data  Z R  41186   0.38 B     15.174 kB
  0-27   data  Z R  37737   0.38 B     13.904 kB
  0-28   data  Z R  37756   0.38 B     13.910 kB
  0-29   data  Z R  34312   0.38 B     12.642 kB
  0-30   data  Z R  34339   0.38 B     12.653 kB
  0-31   data  Z R  30896   0.38 B     11.384 kB
  0-32   data  Z R  30873   0.38 B     11.378 kB
  0-33   data  Z R  27458   0.38 B     10.120 kB
  0-34   data  Z R  27468   0.38 B     10.123 kB
  0-35   data  Z R  27467   0.38 B     10.123 kB
  0-36   data  Z R  24013   0.38 B     8.851 kB
  0-37   data  Z R  24039   0.38 B     8.859 kB
  0-38   data  Z R  24015   0.38 B     8.851 kB
  0-39   data  Z R  20585   0.38 B     7.590 kB
  0-40   data  Z R  20591   0.38 B     7.590 kB
  0-41   data  Z R  20597   0.38 B     7.593 kB
  0-42   data  Z R  20608   0.38 B     7.596 kB
  0-43   data  Z R  17149   0.38 B     6.324 kB
  0-44   data  Z R  17156   0.38 B     6.327 kB
  0-45   data  Z R  17146   0.38 B     6.324 kB
  0-46   data  Z R  17166   0.38 B     6.330 kB
  0-47   data  Z R  13733   0.38 B     5.066 kB
  0-48   data  Z R  13726   0.38 B     5.063 kB
  0-49   data  Z R  13730   0.38 B     5.066 kB
  0-50   data  Z R  13734   0.38 B     5.066 kB
  0-51   data  Z R  13748   0.38 B     5.072 kB
  0-52   data  Z R  10290   0.38 B     3.800 kB
  0-53   data  Z R  10305   0.38 B     3.806 kB
  0-54   data  Z R  10296   0.38 B     3.800 kB
  0-55   data  Z R  10289   0.38 B     3.800 kB
  0-56   data  Z R  10292   0.38 B     3.800 kB
  0-57   data  Z R  10291   0.38 B     3.800 kB
  0-58   data  Z R  10291   0.38 B     3.800 kB
  0-59   data  Z R  6866    0.38 B     2.538 kB
  0-60   data  Z R  6853    0.38 B     2.532 kB
  0-61   data  Z R  6863    0.38 B     2.535 kB
  0-62   data  Z R  6865    0.38 B     2.538 kB
  0-63   data  Z R  6860    0.38 B     2.535 kB
  0-64   data  Z R  6868    0.38 B     2.538 kB
  0-65   data  Z R  6867    0.38 B     2.538 kB
  0-66   data  Z R  6859    0.38 B     2.535 kB
  0-67   data  Z R  6858    0.38 B     2.535 kB
  0-68   data  Z R  6870    0.38 B     2.538 kB
  0-69   data  Z R  6853    0.38 B     2.532 kB
  0-70   data  Z R  3434    0.38 B     1.274 kB
  0-71   data  Z R  3433    0.38 B     1.274 kB
  0-72   data  Z R  3429    0.38 B     1.271 kB
  0-73   data  Z R  3429    0.38 B     1.271 kB
  0-74   data  Z R  3434    0.38 B     1.274 kB
  0-75   data  Z R  3434    0.38 B     1.274 kB
  0-76   data  Z R  3428    0.38 B     1.271 kB
  0-77   data  Z R  3429    0.38 B     1.271 kB
  0-78   data  Z R  3436    0.38 B     1.274 kB
  0-79   data  Z R  3432    0.38 B     1.271 kB
  0-80   data  Z R  3433    0.38 B     1.274 kB
  0-81   data  Z R  3429    0.38 B     1.271 kB
  0-82   data  Z R  3436    0.38 B     1.274 kB
  0-83   data  Z R  3429    0.38 B     1.271 kB
  0-84   data  Z R  3431    0.38 B     1.271 kB
  0-85   data  Z R  3436    0.38 B     1.274 kB
  0-86   data  Z R  3435    0.38 B     1.274 kB
  0-87   data  Z R  3430    0.38 B     1.271 kB
  0-88   data  Z R  3432    0.38 B     1.271 kB
  0-89   data  Z R  3431    0.38 B     1.271 kB
  0-90   data  Z R  3427    0.38 B     1.271 kB
  0-91   data  Z R  3428    0.38 B     1.271 kB
  0-92   data  Z R  3434    0.38 B     1.274 kB
  0-93   data  Z R  3436    0.38 B     1.274 kB
```
Page sizes after
```
Column: shipmode
--------------------------------------------------------------------------------
  page   type  enc  count   avg size   size       rows     nulls   min / max
  0-D    dict  Z _  7       8.29 B     58 B
  0-1    data  Z R  126975  0.38 B     46.755 kB
  0-2    data  Z R  126959  0.38 B     46.749 kB
  0-3    data  Z R  126999  0.38 B     46.764 kB
  0-4    data  Z R  126950  0.38 B     46.746 kB
  0-5    data  Z R  126975  0.38 B     46.755 kB
  0-6    data  Z R  126963  0.38 B     46.752 kB
  0-7    data  Z R  126962  0.38 B     46.752 kB
  0-8    data  Z R  127006  0.38 B     46.767 kB
  0-9    data  Z R  127001  0.38 B     46.767 kB
  0-10   data  Z R  126959  0.38 B     46.749 kB
  0-11   data  Z R  127018  0.38 B     46.773 kB
  0-12   data  Z R  126992  0.38 B     46.761 kB
  0-13   data  Z R  127001  0.38 B     46.767 kB
  0-14   data  Z R  126955  0.38 B     46.749 kB
  0-15   data  Z R  126987  0.38 B     46.761 kB
  0-16   data  Z R  126945  0.38 B     46.746 kB
  0-17   data  Z R  127014  0.38 B     46.771 kB
  0-18   data  Z R  126992  0.38 B     46.761 kB
  0-19   data  Z R  126986  0.38 B     46.761 kB
  0-20   data  Z R  126970  0.38 B     46.755 kB
  0-21   data  Z R  127007  0.38 B     46.767 kB
  0-22   data  Z R  126980  0.38 B     46.758 kB
  0-23   data  Z R  126926  0.38 B     46.737 kB
  0-24   data  Z R  127016  0.38 B     46.771 kB
  0-25   data  Z R  126967  0.38 B     46.752 kB
  0-26   data  Z R  126945  0.38 B     46.746 kB
  0-27   data  Z R  126990  0.38 B     46.761 kB
  0-28   data  Z R  127006  0.38 B     46.767 kB
  0-29   data  Z R  126985  0.38 B     46.761 kB
  0-30   data  Z R  126986  0.38 B     46.761 kB
  0-31   data  Z R  126955  0.38 B     46.749 kB
  0-32   data  Z R  126986  0.38 B     46.761 kB
  0-33   data  Z R  127003  0.38 B     46.767 kB
  0-34   data  Z R  127014  0.38 B     46.771 kB
  0-35   data  Z R  126959  0.38 B     46.749 kB
  0-36   data  Z R  126989  0.38 B     46.761 kB
  0-37   data  Z R  126973  0.38 B     46.755 kB
  0-38   data  Z R  126972  0.38 B     46.755 kB
  0-39   data  Z R  34312   0.38 B     12.642 kB
```

## Release notes

( ) This is not user-visible or docs only and no release notes are required.
( ) Release notes are required, please propose a release note for me.
(x) Release notes are required, with the following suggested text:

```markdown
# Hive
* Fix inefficient packing of data within files written by the optimized parquet writer. ({issue}`17373`)

# Hudi, Delta, Iceberg
* Fix inefficient packing of data within files written by the parquet writer. ({issue}`17373`)
```