@raunaqmorarka 2023-05-08
[Scan after insert parquet snappy optimized writer.pdf](https://github.com/trinodb/trino/files/11420393/Scan.after.insert.parquet.snappy.optimized.writer.pdf)